Move cachingservicetests to caching folder
